+++ Direct@claudiux/files/Direct@claudiux/applet.js
@@ -150,7 +150,6 @@ class RecentFileMenuItem extends IconMenuItem {
 
         super(text, icon);
 
-        //~ this.connect("activate", Lang.bind(this, function(actor, event) {
         this.connect("activate", (actor, event) => {
             let button = event.get_button();
             if (button == 3) {
@@ -160,7 +159,6 @@ class RecentFileMenuItem extends IconMenuItem {
             } else {
                 Gio.app_info_launch_default_for_uri(uri, global.create_app_launch_context());
             }
-        //~ }));
         });
 
         let info = (showUri) ? decodeURIComponent(uri.replace("file://", "").replace(HOME_DIR, "~")) : text;
@@ -382,6 +380,7 @@ class DirectApplet extends Applet.TextIconApplet {
         this.settings.bind("favoriteSizeLimit", "favoriteSizeLimit");
         this.settings.bind("sortingMethod", "sortingMethod");
         this.settings.bind("favoriteSortingMethod", "favoriteSortingMethod");
+        this.settings.bind("favoriteCustomList", "favoriteCustomList");
         this.settings.bind("recentShowUri", "recentShowUri");
         this.settings.bind("favoriteShowUri", "favoriteShowUri");
         this.settings.bind("favoriteIconIsStar", "favoriteIconIsStar");
@@ -424,8 +423,25 @@ class DirectApplet extends Applet.TextIconApplet {
         }
         if ( this.keyOpen.length === 0 ) return;
         this.keyId = "Direct-open";
-        //~ Main.keybindingManager.addHotKey(this.keyId, this.keyOpen, Lang.bind(this, this.openMenu));
         Main.keybindingManager.addHotKey(this.keyId, this.keyOpen, () => this.openMenu());
+    }
+    
+    favoritePopulateList() {
+        var infos = this.favorites.get_favorites(null);
+        var favoriteCustomList = this.favoriteCustomList;
+        var favoriteCustomURIs = [];
+        for (let fav of favoriteCustomList)
+                favoriteCustomURIs.push(fav.uri);
+        infos.forEach( (inf) => {
+            if (favoriteCustomURIs.indexOf(inf.uri) < 0) {
+                favoriteCustomList.push({"menu": true, "name": inf.display_name, "uri": inf.uri});
+                favoriteCustomURIs.push(inf.uri);
+            }
+        });
+        let _to = setTimeout(() => {
+            clearTimeout(_to);
+            this.settings.setValue("favoriteCustomList", favoriteCustomList);
+        }, 2100);
     }
 
     favAppsPopulateList() {
@@ -483,11 +499,11 @@ class DirectApplet extends Applet.TextIconApplet {
         let favAppsPaneBox = new St.BoxLayout({ style_class: "xCenter-pane" });
         let recentPaneBox = new St.BoxLayout({ style_class: "xCenter-pane" });
 
-		let userSettingsImage = new St.Icon({ icon_name: "system-settings", icon_size: 16, icon_type: St.IconType.SYMBOLIC });
-		let systemSettingsImage = new St.Icon({ icon_name: "system-settings", icon_size: 16, icon_type: St.IconType.SYMBOLIC });
-		let favoritesSettingsImage = new St.Icon({ icon_name: "system-settings", icon_size: 16, icon_type: St.IconType.SYMBOLIC });
-		let favAppsSettingsImage = new St.Icon({ icon_name: "system-settings", icon_size: 16, icon_type: St.IconType.SYMBOLIC });
-		let recentSettingsImage = new St.Icon({ icon_name: "system-settings", icon_size: 16, icon_type: St.IconType.SYMBOLIC });
+		let userSettingsImage = new St.Icon({ icon_name: "system-settings-symbolic", icon_size: this.iconSize, icon_type: St.IconType.SYMBOLIC });
+		let systemSettingsImage = new St.Icon({ icon_name: "system-settings-symbolic", icon_size: this.iconSize, icon_type: St.IconType.SYMBOLIC });
+		let favoritesSettingsImage = new St.Icon({ icon_name: "system-settings-symbolic", icon_size: this.iconSize, icon_type: St.IconType.SYMBOLIC });
+		let favAppsSettingsImage = new St.Icon({ icon_name: "system-settings-symbolic", icon_size: this.iconSize, icon_type: St.IconType.SYMBOLIC });
+		let recentSettingsImage = new St.Icon({ icon_name: "system-settings-symbolic", icon_size: this.iconSize, icon_type: St.IconType.SYMBOLIC });
         let userSettingsButton = new St.Button();
         let systemSettingsButton = new St.Button();
         let favoritesSettingsButton = new St.Button();
@@ -543,7 +559,6 @@ class DirectApplet extends Applet.TextIconApplet {
                 userTitle.addActor(userSearchButton);
 
                 userSearchButton.add_actor(userSearchImage);
-                //~ userSearchButton.connect("clicked", Lang.bind(this, this.search, HOME_DIR));
                 userSearchButton.connect("clicked", (a, b) => this.search(a, b, HOME_DIR));
                 new Tooltips.Tooltip(userSearchButton, _("Search Home Folder"));
                 
@@ -584,7 +599,6 @@ class DirectApplet extends Applet.TextIconApplet {
                 systemTitle.addActor(systemSettingsButton);
                 let systemSearchImage = new St.Icon({ icon_name: "edit-find", icon_size: 16, icon_type: St.IconType.SYMBOLIC });
                 systemSearchButton.add_actor(systemSearchImage);
-                //~ systemSearchButton.connect("clicked", Lang.bind(this, this.search));
                 systemSearchButton.connect("clicked", (a, b) => this.search(a, b));
                 new Tooltips.Tooltip(systemSearchButton, _("Search File System"));
                 
@@ -601,8 +615,6 @@ class DirectApplet extends Applet.TextIconApplet {
                 systemPane.actor.add_actor(systemScrollBox);
                 systemScrollBox.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC);
                 let systemVscroll = systemScrollBox.get_vscroll_bar();
-                //~ systemVscroll.connect("scroll-start", Lang.bind(this, function() { this.menu.passEvents = true; }));
-                //~ systemVscroll.connect("scroll-stop", Lang.bind(this, function() { this.menu.passEvents = false; }));
                 systemVscroll.connect("scroll-start", () => { this.menu.passEvents = true; });
                 systemVscroll.connect("scroll-stop", () => { this.menu.passEvents = false; });
 
@@ -634,8 +646,6 @@ class DirectApplet extends Applet.TextIconApplet {
                 favoritesPane.actor.add_actor(favoritesScrollBox);
                 favoritesScrollBox.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC);
                 let favVscroll = favoritesScrollBox.get_vscroll_bar();
-                //~ favVscroll.connect("scroll-start", Lang.bind(this, function() { this.menu.passEvents = true; }));
-                //~ favVscroll.connect("scroll-stop", Lang.bind(this, function() { this.menu.passEvents = false; }));
                 favVscroll.connect("scroll-start", () => { this.menu.passEvents = true; });
                 favVscroll.connect("scroll-stop", () => { this.menu.passEvents = false; });
 
@@ -667,8 +677,6 @@ class DirectApplet extends Applet.TextIconApplet {
                 favAppsPane.actor.add_actor(favAppsScrollBox);
                 favAppsScrollBox.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC);
                 let favAppsVscroll = favAppsScrollBox.get_vscroll_bar();
-                //~ favAppsVscroll.connect("scroll-start", Lang.bind(this, function() { this.menu.passEvents = true; }));
-                //~ favAppsVscroll.connect("scroll-stop", Lang.bind(this, function() { this.menu.passEvents = false; }));
                 favAppsVscroll.connect("scroll-start", () => { this.menu.passEvents = true; });
                 favAppsVscroll.connect("scroll-stop", () => { this.menu.passEvents = false; });
 
@@ -867,24 +875,64 @@ class DirectApplet extends Applet.TextIconApplet {
                     else return 0;
                 }
             );
+        } else if (this.favoriteSortingMethod === "customSort") {
+            var favoriteCustomURIs = [];
+            for (let fav of this.favoriteCustomList)
+                favoriteCustomURIs.push(fav.uri);
+            infos = infos.sort ( (a, b) => {
+                let aURI = a.uri;
+                let bURI = b.uri;
+                if (favoriteCustomURIs.indexOf(aURI) < favoriteCustomURIs.indexOf(bURI)) return -1;
+                if (favoriteCustomURIs.indexOf(aURI) > favoriteCustomURIs.indexOf(bURI)) return 1;
+                return 0;
+            });
         }
 
-        if ( this.favoriteSizeLimit !== 0 && this.favorites.get_n_favorites() > this.favoriteSizeLimit )
+        if ( this.favoriteSortingMethod !== "customSort" && this.favoriteSizeLimit !== 0 && this.favorites.get_n_favorites() > this.favoriteSizeLimit )
             infos = infos.slice( 0, this.favoriteSizeLimit );
 
-        if (infos.length > 0) {
-            for (let i = 0; i < infos.length; i++) {
-                let info = infos[i];
-
-                let button = new FavoriteMenuItem(info, this);
-
-                button.connect("activate", (button, event)=> {
-                    this.favorites.launch(button.info.uri, event.get_time());
-                    this.menu.toggle();
-                })
-
-                this._favoriteButtons.push(button);
-                this.favoriteSection.addMenuItem(button);
+        if (this.favoriteSortingMethod === "customSort") {
+            if (this.favoriteCustomList.length > 0) {
+                for (let i = 0; i < this.favoriteCustomList.length; i++) {
+                    let _info = this.favoriteCustomList[i];
+                    if (_info["menu"] == false) continue;
+    
+                    var info = null;
+                    for (let _inf of infos) {
+                        if (_inf.uri === _info["uri"]) {
+                            info = _inf;
+                            break
+                        }
+                    }
+                    
+                    if (info == null) continue;
+                    
+                    let button = new FavoriteMenuItem(info, this);
+    
+                    button.connect("activate", (button, event)=> {
+                        this.favorites.launch(button.info.uri, event.get_time());
+                        this.menu.toggle();
+                    })
+    
+                    this._favoriteButtons.push(button);
+                    this.favoriteSection.addMenuItem(button);
+                }
+            }
+        } else {
+            if (infos.length > 0) {
+                for (let i = 0; i < infos.length; i++) {
+                    let info = infos[i];
+    
+                    let button = new FavoriteMenuItem(info, this);
+    
+                    button.connect("activate", (button, event)=> {
+                        this.favorites.launch(button.info.uri, event.get_time());
+                        this.menu.toggle();
+                    })
+    
+                    this._favoriteButtons.push(button);
+                    this.favoriteSection.addMenuItem(button);
+                }
             }
         }
     }